requirements.txt
apt-xapian-index==0.47
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
Django==1.11.3
django-filter==1.0.4
django-safedelete==0.4.2
djangorestframework==3.6.3
html5lib==0.999
httplib2==0.9.1
idna==2.0
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Markdown==2.6.8
oauthlib==1.0.3
onboard==1.2.0
oneconf==0.3.9
pexpect==4.0.1
Pillow==3.1.2
piston-mini-client==0.7.5
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
python-apt===1.1.0.b1-elementary2-ubuntu16.04.1
python-debian==0.1.27
python-systemd==231
pytz==2017.2
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
sessioninstaller==0.0.0
six==1.10.0
software-center-aptd-plugins==0.0.0
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
virtualenv==15.0.1
xkit==0.0.0
